# quanto-falta-server/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Tô Contando API
  version: 1.0.0
  description: API central versionada para app Android, site publico, painel administrativo e servicos internos.
servers:
  - url: https://api.tocontando.com.br/api/v1
    description: Production API
  - url: https://api.tocontando.com.br/api/v1
    description: Legacy Production API
  - url: https://api-staging.quantofalta.shop/api/v1
    description: Staging API
  - url: http://localhost:8787/api/v1
    description: Local
security: []
...
paths:
...
  /admin/dashboard:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Retorna resumo administrativo.
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Error"
...
  /admin/devices:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Lista instalacoes/dispositivos.
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /admin/devices/{id}/diagnostics:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Retorna diagnosticos de notificacoes de um dispositivo especifico.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/Error"
  /admin/feedback:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Lista feedbacks.
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /admin/feedback/{id}:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Detalha um feedback.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Atualiza status administrativo de feedback.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
  /admin/ota/releases:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Lista releases OTA.
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Cria uma release OTA.
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
...
  /admin/config:
    patch:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Atualiza configuracao remota permitida.
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
  /admin/settings:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Lista configuracoes administrativas.
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
  /admin/testers:
    get:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Lista testers administrativos.
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      tags: [Admin]
      security:
        - cloudflareAccess: []
      summary: Cria tester.
      responses:
        "201":
          $ref: "#/components/responses/Success"
...
components:
  securitySchemes:
    cloudflareAccess:
      type: http
      scheme: bearer
      bearerFormat: Cloudflare Access JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Success:
      description: Resposta padronizada de sucesso.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
    Error:
      description: Resposta padronizada de erro.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success:
          const: true
        data:
          description: Dados da resposta.
        error:
          type: "null"
        meta:
          $ref: "#/components/schemas/Meta"
    ErrorEnvelope:
      type: object
      required: [success, data, error, meta]
      properties:
        success:
          const: false
        data:
          type: "null"
        error:
          type: object
          required: [code, message]
          properties:
            code:
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - RATE_LIMITED
                - INVALID_TOKEN
                - TOKEN_EXPIRED
                - TOKEN_ALREADY_USED
                - PURCHASE_INVALID
                - PREMIUM_EXPIRED
                - APP_UPDATE_REQUIRED
                - FEATURE_DISABLED
                - INTERNAL_ERROR
            message:
              type: string
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
          format: uuid
...